raftstore: extract a big chunk of snap mod
